Rename fields in createUser, updateUser Activities
